Move silver_measurements schema
