+++ source/core/tests/gold/helpers/silver_builder.py
@@ -2,7 +2,7 @@ import random
 from datetime import datetime
 from decimal import Decimal
 
-from core.gold.domain.schemas.silver_measurements import silver_measurements_schema
+from core.silver.domain.schemas.silver_measurements import silver_measurements_schema
 
 
 class SilverMeasurementsDataFrameBuilder:
